编译bug修复

+++ programs/bumpin-trade/src/state/state.rs
@@ -26,7 +26,7 @@ pub struct State {
     pub redeem_fee_staking_rewards_ratio: u128,
     pub redeem_fee_pool_rewards_ratio: u128,
     pub pool_rewards_interval_limit: u128,
-    pub init_fee: u128,
+    pub init_fee: u64,
     pub open_fee_rate: u128,
     pub staking_fee_reward: FeeReward,
     pub dao_fee_reward: FeeReward,
